feat: initial asciidoc support
Added a parser for asciidoc documents using asciidoctor.js.

The commit also adds the `react-simple-icons` package and
`asciidoctor-highlight.js`, an extension that allows us to
visualise the HTML code.

To make things work our renderer adds the document's scs base url to the
`imagesdir` attribute except when it already is an URL, that way the
images are loaded from the report directory.

The renderer also overrides the background on `.prose pre` adding the
class `prose-pre:bg-stone-400` to avoid the black background when
viewing the code, we probably need to review the themes.

The asciidoctor parser is configured in `safe` mode, which means that the
`includes` do not work.